requirements.txt
pycryptodome==3.21.0
glom==24.11.0
pytest==8.3.4